fixed prepending months, began pr tracking

+++ src/BFN.Data/Models/TrainingLog.cs
@@ -3,7 +3,7 @@ namespace BFN.Data.Models;
 public class TrainingLog : BaseModel
 {
     public int ExerciseId { get; set; }
-    public int MetricWeight { get; set; }
+    public double MetricWeight { get; set; }
     public int Reps { get; set; }
     public int Distance { get; set; }
     public int DurationSeconds { get; set; }
